// src/shared/enum/country-enum.ts
export enum CountryEnum {
  AR = "AR",
  BO = "BO",
  BR = "BR",
  CL = "CL",
  CO = "CO",
  CR = "CR",
  CU = "CU",
  DO = "DO",
  EC = "EC",
  ES = "ES",
  GT = "GT",
  HN = "HN",
  MX = "MX",
  NI = "NI",
  PA = "PA",
  PE = "PE",
  PY = "PY",
  SV = "SV",
  US = "US",
  UY = "UY",
  VE = "VE"
}